- fix Linux 64bit compilation. - move LoadScript() from host.cpp to jsHelper. - fix potential deadlock in host.cpp watchDog management. - jsstd: fix a deadlock in SandboxEval() function. - bug fix in jsiconv with wide string destination. - add QA tests for jsiconv module. - jshost.cpp: add a bootstrap system (-b option) that automatically run (host's name + .js) filename. Remove the argument of the -u option (unsafeMode). - fix jsz Makefile.

+++ src/jsdebug/static.cpp
@@ -787,8 +787,8 @@ DEFINE_FUNCTION( GetObjectPrivate ) {
 		*J_RVAL = JSVAL_VOID;
 		return JS_TRUE;
 	}
-	unsigned int n;
-	n = (unsigned int)JS_GetPrivate(cx, JSVAL_TO_OBJECT( J_ARG( 1 ) ));
+	unsigned long n;
+	n = (unsigned long)JS_GetPrivate(cx, JSVAL_TO_OBJECT( J_ARG( 1 ) ));
 	J_CHK( JS_NewNumberValue(cx, (double)n, J_RVAL) );
 	return JS_TRUE;
 	JL_BAD;
